+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- "quiet" ioption for launcher
+
 ## v2.13.0 - 2023-03-23
 
 ### Fixed